StopRow: Make the component responsive
Also touches and tunes a lot of small padding and margin values to
ensure thinkgs line up nicely with each others on multiple screen
sizes. Style is not based on viewport width, but instead on the
containing elements width.
