+++ ui/src/components/stop-registry/components/StopTableRow/StopTableRow.tsx
@@ -1,6 +1,6 @@
 import { DateTime } from 'luxon';
 import { FC, ReactNode } from 'react';
-import { twMerge } from 'tailwind-merge';
+import { twJoin, twMerge } from 'tailwind-merge';
 import {
   ActionButtonsTd,
   ActionMenuTd,
@@ -10,13 +10,14 @@ import {
   ValidityPeriodTd,
 } from './components';
 import { IconsTd } from './components/IconsTd';
+import { NameAndIconsTd } from './components/NameAndIconsTd';
+import { PlatformNumberTd } from './components/PlatformNumberTd';
 import { SelectRowTd } from './components/SelectRowTd';
 import { StopSearchRow } from './types';
 
 type StopTableRowBaseProps = {
   readonly actionButtons: ReactNode;
   readonly className?: string;
-  readonly inEditMode?: boolean;
   readonly menuItems: ReactNode;
   readonly observationDate: DateTime;
   readonly stop: StopSearchRow;
@@ -26,6 +27,7 @@ export type SelectableStopTableRowProps = {
   readonly selectable: true;
   readonly isSelected: boolean;
   readonly onToggleSelection: (rowId: string) => void;
+  readonly compact?: never;
 };
 
 export type NonSelectableStopTableRowProps = {
@@ -41,12 +43,14 @@ const testIds = {
   row: (label: string) => `StopTableRow::row::${label}`,
 };
 
-const yBorderClassNames = 'border-y border-y-light-grey';
+const yBorder = 'border-y border-y-light-grey';
+const lBorder = 'border-l border-l-background';
+const atWide = '@max-5xl:hidden';
+const atNarrow = '@5xl:hidden';
 
 export const StopTableRow: FC<StopTableRowProps> = ({
   actionButtons,
   className,
-  inEditMode,
   isSelected,
   menuItems,
   observationDate,
@@ -70,37 +74,65 @@ export const StopTableRow: FC<StopTableRowProps> = ({
         />
       )}
 
-      <PriorityTd className={`w-auto ${yBorderClassNames}`} stop={stop} />
+      <PriorityTd className={twJoin(yBorder, 'w-0')} stop={stop} />
 
       <LabelAndTimingPlaceTd
-        className={`w-auto px-4 py-3 pr-20 ${yBorderClassNames}`}
+        className={twJoin(yBorder, 'w-0 p-3 align-top')}
         observationDate={observationDate}
         stop={stop}
       />
 
+      <PlatformNumberTd className={twJoin(yBorder, 'p-3 pl-0')} stop={stop} />
+
       <NameTd
-        className={`w-full px-8 py-3 ${yBorderClassNames} text-sm`}
+        className={twJoin(
+          yBorder,
+          lBorder,
+          atWide,
+          'w-full p-3 align-top text-sm @5xl:px-8',
+        )}
         stop={stop}
       />
 
       <IconsTd
-        className={`w-auto px-2 ${yBorderClassNames} border-l border-l-background whitespace-nowrap`}
+        className={twJoin(
+          yBorder,
+          lBorder,
+          atWide,
+          'w-auto p-3 whitespace-nowrap',
+        )}
+        stop={stop}
+      />
+
+      <NameAndIconsTd
+        className={twJoin(
+          yBorder,
+          lBorder,
+          atNarrow,
+          'w-full p-3 align-top whitespace-nowrap',
+        )}
         stop={stop}
       />
 
       <ValidityPeriodTd
-        className={`w-auto px-8 py-3 ${yBorderClassNames} border-l border-l-background whitespace-nowrap`}
+        className={twJoin(
+          yBorder,
+          lBorder,
+          'w-auto p-3 align-top @5xl:px-8',
+          // Align to the center of the action buttons/menu
+          // calc: (h-10)[action btn height] - 1rem [text height] / 2 + pt-3
+          '@5xl:pt-[calc((((var(--spacing)*10)-1rem)/2)+(var(--spacing)*3))]',
+        )}
         stop={stop}
       />
 
       <ActionButtonsTd
         actionButtons={actionButtons}
-        className={`w-auto py-3 ${yBorderClassNames}`}
+        className={twJoin(yBorder, 'w-0 py-3 pr-1 align-top @max-5xl:hidden')}
       />
 
       <ActionMenuTd
-        className={`w-auto py-3 pr-4 ${inEditMode ? '' : 'pl-4'} ${yBorderClassNames}`}
-        inEditMode={inEditMode}
+        className={twJoin(yBorder, 'w-0 p-3 pr-4 align-top')}
         menuItems={menuItems}
       />
     </tr>
