+++ ui/src/components/stop-registry/components/StopTableRow/components/NameAndIconsTd.tsx
@@ -1,14 +1,18 @@
 import { FC } from 'react';
 import { Column, Row } from '../../../../../layoutComponents';
 import { StopRowTdProps } from '../types';
+import { Icons } from './Icons';
 
-export const NameTd: FC<StopRowTdProps> = ({ className, stop }) => (
+export const NameAndIconsTd: FC<StopRowTdProps> = ({ className, stop }) => (
   <td className={className}>
     <Column>
       <Row className="font-bold">
         {stop.nameFin} | {stop.nameSwe}
       </Row>
       <Row>{stop.description}</Row>
+      <Row>
+        <Icons stop={stop} />
+      </Row>
     </Column>
   </td>
 );
