+++ ui/src/components/stop-registry/components/StopTableRow/components/PlatformNumberTd.tsx
@@ -1,0 +1,9 @@
+import { FC } from 'react';
+import { twMerge } from 'tailwind-merge';
+import { StopRowTdProps } from '../types';
+
+export const PlatformNumberTd: FC<StopRowTdProps> = ({ className, stop }) => (
+  <td className={twMerge('align-text-top font-bold', className)}>
+    {stop.platformNumber}
+  </td>
+);
