+++ ui/src/components/stop-registry/components/StopTableRow/components/ValidityPeriodTd.tsx
@@ -7,11 +7,14 @@ export const ValidityPeriodTd: FC<StopRowTdProps> = ({ className, stop }) => {
   const { t } = useTranslation();
   return (
     <td className={className}>
-      <span title={t(($) => $.accessibility.stops.validityPeriod)}>
-        {mapToShortDate(stop.validityStart)}
-        <span className="mx-1">-</span>
-        {mapToShortDate(stop.validityEnd)}
-      </span>
+      <div
+        className="flex flex-row whitespace-nowrap @max-3xl:flex-col"
+        title={t(($) => $.accessibility.stops.validityPeriod)}
+      >
+        <span>{mapToShortDate(stop.validityStart)}</span>
+        <span className="mx-1 @max-3xl:hidden">-</span>
+        <span>{mapToShortDate(stop.validityEnd)}</span>
+      </div>
     </td>
   );
 };
